// mollys-movies-ui/stubs/api/model/movie-download.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { MovieDownloadStatusCode } from './movie-download-status-code';

/**
 * 
 * @export
 * @interface MovieDownload
 */
export interface MovieDownload {
    /**
     * 
     * @type {string}
     * @memberof MovieDownload
     */
    imdbCode?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MovieDownload
     */
    externalId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof MovieDownload
     */
    name?: string | null;
    /**
     * 
     * @type {MovieDownloadStatusCode}
     * @memberof MovieDownload
     */
    status?: MovieDownloadStatusCode;
}